Replace targeters with test targeting flashpoint.

+++ CORE/src/main/java/com/variant/core/schema/impl/TestsParser.java
@@ -330,7 +330,7 @@ public class TestsParser implements Keywords {
 			}
 		}
 		
-		return new TestExperienceImpl(name, weight.doubleValue(), isControl);	
+		return new TestExperienceImpl(name, weight, isControl);	
 	}
 	
 	/**
